requirements.txt
flask
python-dotenv
flask-wtf

python-shogi
pillow
numpy
imageio
pygifsicle
